requirements.txt
flask
numpy
librosa
tensorflow-cpu
opencv-python
sampleaudio
pydub
ffmpeg